LayoutTests/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-inline-editing-host.tentative_white-space=normal&display=block&method=enter-expected.txt

abcd



FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\"><br><br></span>" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by a <br> (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by a <br> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\"><br><br></span><br>" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span><br>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by text (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by text (defaultParagraphSeparator=div) expected "<span contenteditable=\"\"><br><br></span>abc" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span>abc"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}<br></span> (defaultParagraphSeparator=div) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}<br></span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\"><br><br></span>" but got "<span contenteditable=\"\"><br><p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">[]abcd</span> (defaultParagraphSeparator=div) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">[]abcd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\"><br>abcd</span>" but got "<span contenteditable=\"\"><p><br></p>abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">ab[]cd</span> (defaultParagraphSeparator=div) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">ab[]cd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">ab<br>cd</span>" but got "<span contenteditable=\"\">ab<p>cd</p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">abcd<br><br></span>" but got "<span contenteditable=\"\">abcd<p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">abcd<br><br></span><br>" but got "<span contenteditable=\"\">abcd<p><br></p></span><br>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by text (defaultParagraphSeparator=div) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by text (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">abcd<br><br></span>efg" but got "<span contenteditable=\"\">abcd<p><br></p></span>efg"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\"><br><br></span>" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by a <br> (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by a <br> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\"><br><br></span><br>" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span><br>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by text (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}</span> followed by text (defaultParagraphSeparator=p) expected "<span contenteditable=\"\"><br><br></span>abc" but got "<span contenteditable=\"\"><p><br></p><p><br></p></span>abc"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">{}<br></span> (defaultParagraphSeparator=p) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">{}<br></span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\"><br><br></span>" but got "<span contenteditable=\"\"><br><p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">[]abcd</span> (defaultParagraphSeparator=p) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">[]abcd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\"><br>abcd</span>" but got "<span contenteditable=\"\"><p><br></p>abcd</span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">ab[]cd</span> (defaultParagraphSeparator=p) assert_equals: A <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">ab[]cd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">ab<br>cd</span>" but got "<span contenteditable=\"\">ab<p>cd</p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">abcd<br><br></span>" but got "<span contenteditable=\"\">abcd<p><br></p></span>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">abcd<br><br></span><br>" but got "<span contenteditable=\"\">abcd<p><br></p></span><br>"
FAIL insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by text (defaultParagraphSeparator=p) assert_equals: A <br> and additional <br> should be inserted when insertParagraph in <span contenteditable style="display:block;white-space:normal">abcd[]</span> followed by text (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">abcd<br><br></span>efg" but got "<span contenteditable=\"\">abcd<p><br></p></span>efg"
